+++ packages/browser/src/notifier.ts
@@ -50,11 +50,6 @@ export class Notifier extends BaseNotifier {
       });
     }
 
-    // TODO: deprecated
-    if (this._opt.ignoreWindowError) {
-      opt.instrumentation.onerror = false;
-    }
-
     this._instrument(opt.instrumentation);
   }